+++ src/main/java/com/epam/lowcost/service/implementations/UserServiceImpl.java
@@ -1,7 +1,9 @@
 package com.epam.lowcost.service.implementations;
 
+import com.epam.lowcost.DAO.interfaces.TicketDAO;
 import com.epam.lowcost.DAO.interfaces.UserDAO;
 import com.epam.lowcost.model.User;
+import com.epam.lowcost.service.interfaces.TicketService;
 import com.epam.lowcost.service.interfaces.UserService;
 import org.mindrot.jbcrypt.BCrypt;
 
@@ -10,9 +12,13 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private UserDAO userDAO;
-
+    private TicketService ticketService;
     public UserServiceImpl(UserDAO userDAO) {
         this.userDAO = userDAO;
+    }
+
+    public void setTicketService(TicketService ticketService) {
+        this.ticketService = ticketService;
     }
 
     @Override
@@ -39,7 +45,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public String deleteUser(long userId) {
-        return userDAO.deleteUser(userId);
+        if (ticketService.deleteTicketsByUserId(userId)) {
+            return userDAO.deleteUser(userId);
+        }
+        return null;
     }
 
     @Override
